[FIX] : 대여 및 반납처리과정, URL 변경
