+++ src/main/java/umc/parasol/domain/shop/application/ShopService.java
@@ -3,6 +3,10 @@ package umc.parasol.domain.shop.application;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import umc.parasol.domain.history.domain.History;
+import umc.parasol.domain.history.domain.Process;
+import umc.parasol.domain.history.domain.repository.HistoryRepository;
+import umc.parasol.domain.history.dto.HistoryRes;
 import umc.parasol.domain.image.domain.Image;
 import umc.parasol.domain.image.domain.repository.ImageRepository;
 import umc.parasol.domain.image.dto.ImageRes;
@@ -11,11 +15,15 @@ import umc.parasol.domain.member.domain.repository.MemberRepository;
 import umc.parasol.domain.shop.domain.Shop;
 import umc.parasol.domain.shop.domain.repository.ShopRepository;
 import umc.parasol.domain.shop.dto.*;
+import umc.parasol.domain.umbrella.application.UmbrellaService;
 import umc.parasol.domain.umbrella.domain.Umbrella;
 import umc.parasol.domain.umbrella.domain.repository.UmbrellaRepository;
+import umc.parasol.global.config.security.token.CurrentUser;
 import umc.parasol.global.config.security.token.UserPrincipal;
+import umc.parasol.global.payload.ApiResponse;
 
 import javax.swing.text.html.Option;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -31,6 +39,11 @@ public class ShopService {
     private final ImageRepository imageRepository;
 
     private final UmbrellaRepository umbrellaRepository;
+
+    private final HistoryRepository historyRepository;
+
+    private final UmbrellaService umbrellaService;
+
 
     /**
      * 매장 리스트 조회
@@ -184,6 +197,88 @@ public class ShopService {
         return createShopRes(updatedShop, imageResList);
 
 
+    }
+
+    /**
+     * 우산 대여 처리
+     * @param user api 호출하는 사용자 객체 (점주)
+     * @param memberId 대여자 객체
+     */
+    @Transactional
+    public ApiResponse rentalUmbrella(@CurrentUser UserPrincipal user, Long memberId) {
+
+        Member owner = findValidMember(user.getId()); //점주
+        Shop targetShop = findValidShopForOwner(owner); //점주의 shop
+
+        Member member = findValidMember(memberId); //대여자
+
+        History history = rentalUmbrella(targetShop, member);
+        historyRepository.save(history);
+
+        HistoryRes record = makeHistoryRes(member, history, null);
+        return new ApiResponse(true, record);
+    }
+
+    /**
+     * 우산 반납 처리
+     * @param user api 호출하는 사용자 객체 (점주)
+     * @param memberId 대여자 객체
+     */
+    @Transactional
+    public ApiResponse returnUmbrella(@CurrentUser UserPrincipal user, Long memberId) {
+        Member owner = findValidMember(user.getId()); //점주
+        Shop targetShop = findValidShopForOwner(owner); //점주의 shop
+
+        Member member = findValidMember(memberId); //대여자
+
+        List<History> remainHistoryList = historyRepository.findAllByMemberOrderByCreatedAtDesc(member)
+                .stream()
+                .filter(history -> history.getProcess() != Process.CLEAR).toList();
+
+        if (remainHistoryList.isEmpty())
+            throw new IllegalStateException("처리되지 않은 우산 내역이 없습니다.");
+
+        History targetHistory = remainHistoryList.get(remainHistoryList.size() - 1);
+        Umbrella targetUmbrella = targetHistory.getUmbrella();
+        targetHistory.updateProcess(Process.CLEAR); // History CLEAR 상태로 변경
+
+        Shop originShop = targetHistory.getFromShop();
+
+        targetUmbrella.updateAvailable(true);
+
+        if (originShop != targetShop) { // 빌렸던 Shop과 다르다면
+            targetUmbrella.updateShop(targetShop);
+        }
+        targetHistory.updateClearedAt(LocalDateTime.now());
+        targetHistory.updateEndShop(targetShop);
+        HistoryRes record = makeHistoryRes(member, targetHistory, targetShop);
+        return new ApiResponse(true, record);
+    }
+
+    private HistoryRes makeHistoryRes(Member member, History history, Shop endShop) {
+        return HistoryRes.builder()
+                .member(member.getNickname())
+                .fromShop(history.getFromShop().getName())
+                .endShop((endShop == null) ?
+                        (history.getEndShop() == null
+                                ? ""
+                                : history.getEndShop().getName())
+                        : endShop.getName())
+                .createdAt(history.getCreatedAt())
+                .clearedAt(history.getClearedAt() == null ? LocalDateTime.of(9999, 12, 31, 23, 59) : history.getClearedAt())
+                .process(history.getProcess())
+                .build();
+    }
+
+    private History rentalUmbrella(Shop targetShop, Member member) {
+        return History.builder()
+                .cost(0)
+                .process(Process.USE)
+                .fromShop(targetShop)
+                .endShop(null)
+                .member(member)
+                .umbrella(umbrellaService.getAnyFreeUmbrella(targetShop))
+                .build();
     }
 
 }
